+++ mslicer/src/render/slice_preview/mod.rs
@@ -13,9 +13,9 @@ pub struct SlicePreviewRenderResources {
 pub struct SlicePreviewRenderCallback {
     pub dimensions: Vector2<u32>,
     pub offset: Vector2<f32>,
+    pub scale: Vector2<f32>,
     pub aspect: f32,
     pub pixel_aspect: f32,
-    pub scale: f32,
 
     pub new_preview: Option<(Vec<u8>, Vec<u8>)>,
 }